+++ taxmate/uae_vat/report/uae_group_vat_status/uae_group_vat_status.js
@@ -1,0 +1,6 @@
+// Copyright (c) 2026, Sourav Singh and contributors
+// For license information, please see license.txt
+
+frappe.query_reports["UAE Group VAT Status"] = {
+	filters: [],
+};